Enemies, shooying, UI, points, death
Enemies will spawn within the map and follow the player. When touching enemies, the player loses a life. If the player has 0 or less hp - he dies and a restart button appears. There is a shot on the left mouse button, each bullet takes away a certain number of lives from the enemy. When an enemy dies - the player gets a point and a chance to get a better weapon that drops from enemies.